+++ Source/sync.h
@@ -10,7 +10,7 @@
 
 namespace devilution {
 
-uint32_t sync_all_monsters(std::byte *pbBuf, uint32_t dwMaxLen);
+size_t sync_all_monsters(std::byte *pbBuf, size_t dwMaxLen);
 uint32_t OnSyncData(const TCmd *pCmd, size_t pnum);
 void sync_init();
